refactor: extension.ts
